--- fixtures/asset-config/package.json
@@ -1,24 +1,0 @@
-{
-	"name": "workers-assets-config-test",
-	"private": true,
-	"scripts": {
-		"dev": "npx wrangler dev",
-		"test:ci": "run-script-os",
-		"test:ci:default": "true",
-		"test:ci:nix": "vitest run",
-		"test:watch": "vitest",
-		"type:tests": "tsc -p ./tsconfig.json"
-	},
-	"devDependencies": {
-		"@cloudflare/vitest-pool-workers": "workspace:*",
-		"@cloudflare/workers-tsconfig": "workspace:*",
-		"@cloudflare/workers-types": "^4.20241106.0",
-		"run-script-os": "^1.1.6",
-		"undici": "catalog:default",
-		"vitest": "catalog:default",
-		"wrangler": "workspace:*"
-	},
-	"volta": {
-		"extends": "../../package.json"
-	}
-}
