+++ packages/workers-shared/asset-worker/tests/test-cases/test-case.ts
@@ -1,0 +1,7 @@
+export type TestCase = {
+	title: string;
+	files: string[];
+	requestPath: string;
+	matchedFile?: string;
+	finalPath?: string;
+};
