+++ packages/workers-shared/package.json
@@ -29,16 +29,17 @@
 		"bundle:router-worker": "esbuild router-worker/src/index.ts --format=esm --bundle --outfile=dist/router-worker.mjs --sourcemap=external",
 		"bundle:router-worker:prod": "pnpm run bundle:router-worker --minify",
 		"check:lint": "eslint . --max-warnings=0",
-		"check:type": "tsc -p ./tsconfig.json && tsc -p ./tsconfig.node.json",
+		"check:type": "pnpm run check:type:package && pnpm run check:type:asset-worker && pnpm run check:type:router-worker",
 		"check:type:asset-worker": "tsc -p ./asset-worker/tsconfig.json && tsc -p ./asset-worker/tests/tsconfig.json",
+		"check:type:package": "tsc -p ./tsconfig.json && tsc -p ./tsconfig.node.json",
 		"check:type:router-worker": "tsc -p ./router-worker/tsconfig.json && tsc -p ./router-worker/tests/tsconfig.json",
 		"clean": "rimraf dist",
 		"deploy": "pnpm run deploy:router-worker && pnpm run deploy:asset-worker",
 		"deploy:asset-worker": "CLOUDFLARE_API_TOKEN=$WORKERS_DEPLOY_AND_CONFIG_CLOUDFLARE_API_TOKEN pnpx wrangler versions upload --experimental-versions -c asset-worker/wrangler.toml",
 		"deploy:router-worker": "CLOUDFLARE_API_TOKEN=$WORKERS_DEPLOY_AND_CONFIG_CLOUDFLARE_API_TOKEN pnpx wrangler versions upload --experimental-versions -c router-worker/wrangler.toml",
 		"dev": "pnpm run clean && concurrently -n bundle:asset-worker,bundle:router-worker -c blue,magenta \"pnpm run bundle:asset-worker --watch\" \"pnpm run bundle:router-worker --watch\"",
-		"test": "vitest",
-		"test:ci": "vitest run",
+		"test": "vitest -c asset-worker/vitest.config.mts",
+		"test:ci": "vitest run -c asset-worker/vitest.config.mts",
 		"types:emit": "tsc index.ts --declaration --emitDeclarationOnly --declarationDir ./dist"
 	},
 	"dependencies": {
