+++ packages/workers-shared/tsconfig.json
@@ -4,7 +4,7 @@
 		"lib": ["es2021"],
 		"module": "es2022",
 		"moduleResolution": "Bundler",
-		"types": ["@cloudflare/workers-types/experimental"],
+		"types": ["@cloudflare/workers-types/experimental", "@types/node"],
 		"noEmit": true,
 		"isolatedModules": true,
 		"allowSyntheticDefaultImports": true,
@@ -12,6 +12,6 @@
 		"strict": true,
 		"skipLibCheck": true
 	},
-	"include": ["**/*.ts"],
+	"include": ["utils/*.ts"],
 	"exclude": ["scripts/**"]
 }
